% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{atc_antineoplastic_id}
\alias{atc_antineoplastic_id}
\title{ATC Antineoplastic Id}
\format{
A vector of length 1
\describe{
\item{\code{concept_id}}{integer}
}
}
\usage{
atc_antineoplastic_id
}
\description{
Concept Id of the ATC Antineoplastics Class
}
\details{
DETAILS
}
\keyword{datasets}
