% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildIngredientExposuresTable.R
\name{buildIngredientExposuresTable}
\alias{buildIngredientExposuresTable}
\title{Build Ingredient Exposures Tables}
\usage{
buildIngredientExposuresTable(
  conn,
  cdmDatabaseSchema,
  writeDatabaseSchema,
  drugExposureIngredientTable
)
}
\arguments{
\item{conn}{Connection object to the database engine}

\item{cdmDatabaseSchema}{Source CDM Drug Exposure Table schema}

\item{writeDatabaseSchema}{Schema where the OncoRegimenFinder tables will be written to.}

\item{drugExposureIngredientTable}{Name of the new Ingredient Exposure Table}
}
\value{
Ingredient Exposure Table in \code{writeDatabaseSchema} that contains all the original fields from the Drug Exposure Table with the addition of an Ingedient Conceot Id column that maps the drug to its RxNorm Ingredient.
}
\description{
This function builds the Ingredient Exposures Table for the algorithm to derive the antineoplastic combinations based on the RxNorm Ingredient of the drug in question.
}
\details{
If the Cohort, Regimen, and Regimen Staging Tables already exist, these tables are renamed with the system date appended in the format "{_YYYY_MM_DD}". The renaming is skipped and the existing Cohort, Regimen and Regimen Staging Tables are dropped if the dated table already exists, indicating that the OncoRegimenFinder was already run once for that system date.
}
\seealso{
\code{\link[OncoRegimenFinder]{atc_antineoplastic_id}},\code{\link[OncoRegimenFinder]{hemonc_classes}},\code{\link[OncoRegimenFinder]{falsepositives}}
\code{\link[pg13]{lsTables}},\code{\link[pg13]{renameTable}},\code{\link[pg13]{appendDate}},\code{\link[pg13]{execute}}
\code{\link[SqlRender]{render}},\code{\link[SqlRender]{readSql}}
}
