% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRegimenIngrTable.R
\name{createRegimenIngrTable}
\alias{createRegimenIngrTable}
\title{Create Regimen Ingredient Table}
\usage{
createRegimenIngrTable(
  conn,
  writeDatabaseSchema,
  cohortTable,
  regimenTable,
  regimenIngredientTable,
  vocabularyTable
)
}
\description{
Create Regimen Ingredient Table
}
