% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createVocabTable.R
\name{createVocabTable}
\alias{createVocabTable}
\title{Process Regimen Table}
\usage{
createVocabTable(
  conn,
  writeDatabaseSchema,
  cdmDatabaseSchema,
  vocabularyTable,
  renameCurrentTables = TRUE
)
}
\description{
Process Regimen Table
}
