% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{falsepositives}
\alias{falsepositives}
\title{False Positive Cancer Drugs}
\format{
A data frame with 283 rows and 1 variables:
\describe{
\item{\code{concept_id}}{character RxNorm Ingredient Concept Id}
}
}
\usage{
falsepositives
}
\description{
Concept Ids that fall under a Cancer-related drug classification, but are unlikely to be used primarily for a cancer.
}
\details{
DETAILS
}
\keyword{datasets}
