% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hemonc_classes}
\alias{hemonc_classes}
\title{HemOnc Antineoplastic Classes}
\format{
A vector
\describe{
\item{\code{concept_id}}{integer}
}
}
\usage{
hemonc_classes
}
\description{
Concept Ids of  HemOnc Antineoplastic Drug Classes
}
\details{
DETAILS
}
\keyword{datasets}
