% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_connDetails.R
\name{make_connDetails}
\alias{make_connDetails}
\title{Make connectionDetails}
\usage{
make_connDetails(dbms = "postgresql", server, user, password, port)
}
\value{
connectionDetail object
}
\description{
Wrapper around the DatabaseConnector::createConnectionDetails function.
}
