% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processRegimenTable.R
\name{processRegimenTable}
\alias{processRegimenTable}
\title{Process Regimen Table}
\usage{
processRegimenTable(
  conn,
  writeDatabaseSchema,
  regimenTable,
  date_lag_input = 30,
  regimen_repeats = 5
)
}
\description{
Process Regimen Table
}
