% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowCountHistory.R
\name{rowCountHistory}
\alias{rowCountHistory}
\title{Row Counts}
\usage{
rowCountHistory(
  conn,
  writeDatabaseSchema,
  vocabularyTable,
  drugExposureIngredientTable,
  cohortTable,
  regimenTable,
  regimenIngredientTable
)
}
\arguments{
\item{conn}{Database connection object}

\item{writeDatabaseSchema}{Schema where tables are written}

\item{vocabularyTable}{Vocabulary Table}

\item{drugExposureIngredientTable}{Drug Exposure Table}

\item{cohortTable}{Cohort Table}

\item{regimenTable}{Regimen Table}

\item{regimenIngredientTable}{Regimen Ingredient Table}
}
\value{
A named list of data frames with counts for each respective argument.
}
\description{
This function takes the names of each of the table arguments for \code{\link{runORF}}, finds all the tables matching the name with or without the appended date in the format "YYYY_mm_dd".
}
\details{
DETAILS
}
\seealso{
\code{\link[rubix]{map_names_set}}
\code{\link[pg13]{lsTables}},\code{\link[pg13]{query}},\code{\link[pg13]{renderRowCount}}
\code{\link[purrr]{map2}},\code{\link[purrr]{set_names}},\code{\link[purrr]{map}}
\code{\link[dplyr]{bind}},\code{\link[dplyr]{select_all}}
}
