% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{topclasses}
\alias{topclasses}
\title{Top Classes}
\format{
A data frame with 99 rows and 4 variables:
\describe{
\item{\code{Vocabulary}}{character Vocabulary Id}
\item{\code{Category}}{character Category assigned}
\item{\code{cancer_drug_category_concept_id}}{character Concept Id for the Category}
\item{\code{cancer_drug_category_concept_name}}{character Conecpt Name for the Category}
}
}
\usage{
topclasses
}
\description{
Top OMOP Classifications for Cancer Drugs
}
\details{
DETAILS
}
\keyword{datasets}
