% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildCohortRegimenTable.R
\name{buildCohortRegimenTable}
\alias{buildCohortRegimenTable}
\title{Build the Cohort and Regimen Tables}
\usage{
buildCohortRegimenTable(
  conn,
  cdmDatabaseSchema,
  cohortDefinitionId = NULL,
  cohortDefinitionIdSchema = NULL,
  writeDatabaseSchema,
  cohortTable,
  regimenTable,
  renameCurrentTables = TRUE,
  drug_classification_id_input
)
}
\arguments{
\item{cohortDefinitionId}{Optional. Cohort Definition Id in the \code{cdmResultSchema}}

\item{cohortDefinitionIdSchema}{Optional. Schema where the the cohort and cohort definition ID point to.}
}
\value{
The Cohort and Regimen Tables in the writeDatabaseSchema
}
\description{
Build the Cohort and Regimen Tables
}
